feat: model 영역 User Class 추가 - UserTest 작성
