+++ src/test/java/model/LottoTest.java
@@ -1,6 +1,5 @@
 package model;
 
-import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
@@ -9,18 +8,6 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class LottoTest {
-    @Test
-    void 랜덤_숫자_로또를_생성하는_경우() {
-        //given
-        int count = 5;
-
-        // when
-        List<Lotto> userLotto = Lotto.getUserLotto(count);
-
-        //then
-        assertThat(userLotto.size()).isEqualTo(5);
-    }
-
     @Test
     void 보너스_숫자가_포함되는_경우() {
         //given
@@ -53,18 +40,6 @@ public class LottoTest {
 
         //then
         assertThat(result).isEqualTo(false);
-    }
-
-    @Test
-    void 사용자가_구입한_로또_갯수를_구하는_경우() {
-        //given
-        int cost = 5000;
-
-        //when
-        int amount = Lotto.getCount(cost);
-
-        //then
-        AssertionsForClassTypes.assertThat(amount).isEqualTo(5);
     }
 
     @Test
